IMPORTANT: Update to how DC/DCTERMS metadata is handled in HTML5
These changes are necessary to produce valid HTML that embeds
Dublin Core metadata. The Kairos style guide ought to be updated
accordingly.

Note that the rel="" syntax for pointing to schema is gone, replaced
by the prefix="" attribute on <head>.

Note also that scheme and language attributes are no longer necessary
on any of the name/content pairs in the <meta> elements. When a
language is specified, it should be done so according to ISO 639 language
codes (e.g. en for English).